+++ DoubleValueDisplayElement.cs
@@ -7,7 +7,7 @@ using MetaphysicsIndustries.Utilities;
 namespace MetaphysicsIndustries.Amethyst
 {
     [Serializable]
-    public class DoubleValueDisplayElement : ValueDisplayElement<double>
+    public class DoubleValueDisplayElement : ValueDisplayElement<float>
     {
         public DoubleValueDisplayElement()
             : base(new ValueDisplayNode("Double Value"), new SizeV(80, 40))
